+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [1.0.4] - 2023-09-16 - @tzoug
+### Enhancement
+- Implemented settings
+    - Option to surround `FLOAT` variables with single quotes. Useful for very large numbers.
+    - Option to replace the variables with their value in the query.
+
 ## [1.0.3] - 2020-06-29 - @gerneio
 ### Enhancement
 - Support converting multiple sp_executesql statements [#6](https://github.com/PejmanNik/sqlops-spexecutesql-to-sql/pull/6)